Inline `Index` class into `LighthouseRunner`
