+++ scanner/lib/lighthouse_runner.rb
@@ -1,4 +1,3 @@
-require 'lib/index'
 require 'lib/scan_output'
 require 'uri'
 require_relative '../../lib/utils'
@@ -58,3 +57,19 @@ class LighthouseRunner
 end
 
 Page = Struct.new :framework_name, :name, :url
+
+class Index
+  def initialize
+    @urlmap = {}
+  end
+
+  def add(framework, page)
+    @urlmap[framework] ||= []
+    @urlmap[framework].push page
+  end
+
+  def generate_index(scan_dir)
+    log :info, 'Generating index file'
+    File.write(scan_dir / 'index.json', JSON.dump(@urlmap))
+  end
+end
